Move mockHttp into a __mocks__ folder
